backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.9
python-dotenv==1.1.1
python-multipart==0.0.20
google-genai==1.35.0
